Add toggling of logging to rosout
Added a button that toggles the logging of messages to the rosout
table. Click on the button first to pause logging, then click on it
again to resume. It will resume logging from the next most recent
message; all messages right after pausing and right before resuming
will be lost.

Also added padding and centering to the compass arrows to make
things look just a tad better.